Move softreboot tests under vmi lifecycle
